# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mapa OSC Data API",
        "description": "É uma API REST que alimenta o website https://www.mapaosc.ipea.gov.br/",
        "contact": {
            "name": "Mapa das OSCs",
            "email": "mapaosc@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/osc/{osc_id}": {
            "get": {
                "tags": [
                    "Osc"
                ],
                "operationId": "get",
                "parameters": [
                    {
                        "name": "osc_id",
                        "in": "path",
                        "description": "Número de identificação da OSC",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma OSC de acordo com o ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Osc"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/osc/projetos/{id_osc}": {
            "get": {
                "tags": [
                    "Projeto"
                ],
                "operationId": "getProjetos",
                "parameters": [
                    {
                        "name": "id_osc",
                        "in": "path",
                        "description": "Número de identificação da OSC",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados e todos os projetos de acordo com a OSC informado."
                    }
                }
            }
        },
...
        "/api/osc/projeto/{id}": {
            "get": {
                "tags": [
                    "Projeto"
                ],
                "operationId": "get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Número de identificação do projeto",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados do projeto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Projeto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Osc": {
                "description": "Osc model",
                "properties": {
                    "id_osc": {
                        "description": "Número de identificação da Osc",
                        "type": "integer"
                    },
                    "tx_apelido_osc": {
                        "description": "Apelido da Osc",
                        "type": "string"
                    },
                    "ft_apelido_osc": {
                        "description": "Fonte de origem do Apelido da Osc",
                        "type": "string"
                    },
                    "cd_identificador_osc": {
                        "description": "CNPJ da Osc",
                        "type": "integer"
                    },
                    "ft_identificador_osc": {
                        "description": "Fonte da origem do CNPJ da Osc",
                        "type": "string"
                    },
                    "bo_osc_ativa": {
                        "description": "Campo descreve ativação da OSC no Portal",
                        "type": "integer"
                    },
                    "bo_nao_possui_projeto": {
                        "description": "Campo que diz se a OSC tem projeto ou não",
                        "type": "integer"
                    },
                    "ft_nao_possui_projeto": {
                        "description": "Fonte da origem do Campo que diz se a OSC tem projeto ou não",
                        "type": "string"
                    },
                    "bo_nao_possui_ps_conselhos": {
                        "description": "Campo que diz se a OSC tem conselhos ou não",
                        "type": "integer"
                    },
                    "ft_nao_possui_ps_conselhos": {
                        "description": "Fonte da origem do Campo que diz se a OSC tem conselhos ou não",
                        "type": "string"
                    },
                    "bo_nao_possui_ps_conferencias": {
                        "description": "Campo que diz se a OSC tem conferencias ou não",
                        "type": "integer"
                    },
                    "ft_nao_possui_ps_conferencias": {
                        "description": "Fonte da origem do Campo que diz se a OSC tem conferencias ou não",
                        "type": "string"
                    },
                    "bo_nao_possui_ps_outros_espacos": {
                        "description": "Campo que diz se a OSC tem outros espaços ou não",
                        "type": "integer"
                    },
                    "ft_nao_possui_ps_outros_espacos": {
                        "description": "Fonte da origem do Campo que diz se a OSC tem outros espaços ou não",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Projeto": {
                "description": "Projeto model",
                "properties": {
                    "id": {
                        "description": "Número de identificação do projeto",
                        "type": "string",
                        "example": ""
                    },
                    "id_osc": {
                        "description": "Identificação da OSC.",
                        "type": "int"
                    },
                    "cd_status_projeto": {
                        "description": "Identificação de status do projeto.",
                        "type": "int"
                    },
                    "cd_abrangencia_projeto": {
                        "description": "Identificação da abrangencia.",
                        "type": "int"
                    },
                    "cd_zona_atuacao_projeto": {
                        "description": "Identificação da zona de atuação.",
                        "type": "int"
                    },
                    "cd_municipio": {
                        "description": "Identificação do município.",
                        "type": "int"
                    },
                    "cd_uf": {
                        "description": "Identificação do estado.",
                        "type": "string"
                    },
                    "tx_nome_projeto": {
                        "description": "Nome do projeto.",
                        "type": "string"
                    },
                    "ft_nome_projeto": {
                        "description": "Fonte do dado sobre o nome do projeto.",
                        "type": "string"
                    },
                    "ft_status_projeto": {
                        "description": "Fonte do dado sobre o status do projeto.",
                        "type": "string"
                    },
                    "dt_data_inicio_projeto": {
                        "description": "Data de inicio do projeto.",
                        "type": "string"
                    },
                    "ft_data_inicio_projeto": {
                        "description": "Fonte do dado sobre data de inicio do projeto.",
                        "type": "string"
                    },
                    "dt_data_fim_projeto": {
                        "description": "Data de fim do projeto.",
                        "type": "string"
                    },
                    "ft_data_fim_projeto": {
                        "description": "Fonte do dado sobre data de inicio do projetoIdentificação da OSC.",
                        "type": "string"
                    },
                    "tx_link_projeto": {
                        "description": "Link do projeto.",
                        "type": "string"
                    },
                    "ft_link_projeto": {
                        "description": "Fonte do dado sobre link do projeto.",
                        "type": "string"
                    },
                    "nr_total_beneficiarios": {
                        "description": "Total de beneficiários.",
                        "type": "string"
                    },
                    "ft_total_beneficiarios": {
                        "description": "Fonte do dado sobre total de beneficiários.",
                        "type": "string"
                    },
                    "nr_valor_captado_projeto": {
                        "description": "Valor captado do projeto.",
                        "type": "string"
                    },
                    "ft_valor_captado_projeto": {
                        "description": "Fonte do dado sobre valor captado do projeto.",
                        "type": "string"
                    },
                    "nr_valor_total_projeto": {
                        "description": "Valor total do projeto.",
                        "type": "string"
                    },
                    "ft_valor_total_projeto": {
                        "description": "Fonte do dado sobre valor total do projeto.",
                        "type": "string"
                    },
                    "ft_abrangencia_projeto": {
                        "description": "Fonte do dado sobre abrangencia.",
                        "type": "string"
                    },
                    "ft_zona_atuacao_projeto": {
                        "description": "Fonte do dado sobre zona de atuação.",
                        "type": "string"
                    },
                    "tx_descricao_projeto": {
                        "description": "Descrição do projeto.",
                        "type": "string"
                    },
                    "ft_descricao_projeto": {
                        "description": "Fonte do dado sobre descrição do projeto.",
                        "type": "string"
                    },
                    "ft_metodologia_monitoramento": {
                        "description": "Fonte do dado sobre metodologia de monitoramento.",
                        "type": "string"
                    },
                    "tx_metodologia_monitoramento": {
                        "description": "Metodologia de monitoramento.",
                        "type": "string"
                    },
                    "tx_identificador_projeto_externo": {
                        "description": "Identificador do projeto externo.",
                        "type": "string"
                    },
                    "ft_identificador_projeto_externo": {
                        "description": "Fonte do dado sobre identificador do projeto externo.",
                        "type": "string"
                    },
                    "bo_oficial": {
                        "description": "BO oficial.",
                        "type": "boolean"
                    },
                    "tx_status_projeto_outro": {
                        "description": "Status de outro projeto.",
                        "type": "string"
                    },
                    "ft_municipio": {
                        "description": "Fonte do dado sobre município.",
                        "type": "string"
                    },
                    "ft_uf": {
                        "description": "Fonte do dado sobre estado.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}